refactor: repository 구분 잘못한 거 수정
